+++ egov/egov-councilweb/src/main/resources/messages/errors.properties
@@ -85,3 +85,5 @@ err.agenda.sequence = Please enter Agenda sequence number greator than last used
 err.resolution.sequence =Please enter Resolution sequence number greator than last used number
 err.meeting.sequence=Please enter Meeting sequence number greator than last used number
 err.agenda.alreadyexists=Agenda number already exists
+err.meeting.no.alreadyexists=Meeting number already exists
+
